Move models/generated to generated/models
