+++ ksmt-runner/src/main/kotlin/org/ksmt/solver/runner/KSolverRunnerExecutor.kt
@@ -8,15 +8,15 @@ import org.ksmt.decl.KConstDecl
 import org.ksmt.decl.KDecl
 import org.ksmt.expr.KExpr
 import org.ksmt.runner.core.KsmtWorkerSession
-import org.ksmt.runner.models.generated.AssertAndTrackParams
-import org.ksmt.runner.models.generated.AssertParams
-import org.ksmt.runner.models.generated.CheckParams
-import org.ksmt.runner.models.generated.CheckWithAssumptionsParams
-import org.ksmt.runner.models.generated.CreateSolverParams
-import org.ksmt.runner.models.generated.PopParams
-import org.ksmt.runner.models.generated.SolverConfigurationParam
-import org.ksmt.runner.models.generated.SolverProtocolModel
-import org.ksmt.runner.models.generated.SolverType
+import org.ksmt.runner.generated.models.AssertAndTrackParams
+import org.ksmt.runner.generated.models.AssertParams
+import org.ksmt.runner.generated.models.CheckParams
+import org.ksmt.runner.generated.models.CheckWithAssumptionsParams
+import org.ksmt.runner.generated.models.CreateSolverParams
+import org.ksmt.runner.generated.models.PopParams
+import org.ksmt.runner.generated.models.SolverConfigurationParam
+import org.ksmt.runner.generated.models.SolverProtocolModel
+import org.ksmt.runner.generated.models.SolverType
 import org.ksmt.solver.KModel
 import org.ksmt.solver.KSolverException
 import org.ksmt.solver.KSolverStatus
