+++ ksmt-runner/src/main/kotlin/org/ksmt/solver/runner/KSolverWorker.kt
@@ -3,8 +3,8 @@ package org.ksmt.solver.runner
 import com.jetbrains.rd.framework.IProtocol
 import org.ksmt.runner.core.KsmtWorkerBase
 import org.ksmt.runner.core.RdServer
-import org.ksmt.runner.models.generated.SolverProtocolModel
-import org.ksmt.runner.models.generated.solverProtocolModel
+import org.ksmt.runner.generated.models.SolverProtocolModel
+import org.ksmt.runner.generated.models.solverProtocolModel
 
 class KSolverWorker(id: Int, process: RdServer) : KsmtWorkerBase<SolverProtocolModel>(id, process) {
     override fun initProtocolModel(protocol: IProtocol): SolverProtocolModel =
