+++ ksmt-runner/src/main/kotlin/org/ksmt/solver/runner/KSolverWorkerProcess.kt
@@ -6,15 +6,15 @@ import org.ksmt.decl.KConstDecl
 import org.ksmt.expr.KExpr
 import org.ksmt.runner.core.ChildProcessBase
 import org.ksmt.runner.core.KsmtWorkerArgs
-import org.ksmt.runner.models.generated.CheckResult
-import org.ksmt.runner.models.generated.ModelEntry
-import org.ksmt.runner.models.generated.ModelFuncInterpEntry
-import org.ksmt.runner.models.generated.ModelResult
-import org.ksmt.runner.models.generated.ModelUninterpretedSortUniverse
-import org.ksmt.runner.models.generated.ReasonUnknownResult
-import org.ksmt.runner.models.generated.SolverProtocolModel
-import org.ksmt.runner.models.generated.UnsatCoreResult
-import org.ksmt.runner.models.generated.solverProtocolModel
+import org.ksmt.runner.generated.models.CheckResult
+import org.ksmt.runner.generated.models.ModelEntry
+import org.ksmt.runner.generated.models.ModelFuncInterpEntry
+import org.ksmt.runner.generated.models.ModelResult
+import org.ksmt.runner.generated.models.ModelUninterpretedSortUniverse
+import org.ksmt.runner.generated.models.ReasonUnknownResult
+import org.ksmt.runner.generated.models.SolverProtocolModel
+import org.ksmt.runner.generated.models.UnsatCoreResult
+import org.ksmt.runner.generated.models.solverProtocolModel
 import org.ksmt.runner.serializer.AstSerializationCtx
 import org.ksmt.solver.KSolver
 import org.ksmt.sort.KBoolSort
