+++ ksmt-runner/src/test/kotlin/org/ksmt/solver/runner/SolverRunnerTest.kt
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.ksmt.KContext
-import org.ksmt.runner.models.generated.SolverType
+import org.ksmt.runner.generated.models.SolverType
 import org.ksmt.solver.KSolver
 import org.ksmt.solver.KSolverStatus
 import org.ksmt.solver.z3.KZ3Solver
